Say the mod doesn't work in creative mode
I and a few others tested this mod in creative mode and was confused as to why this didn't work.
The showcase video showcases this feature, but as it turns out, not a lot of people watches the video. This change highlights possibly the most important note from that video, since you can't trust everybody to watch that video.